requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
install==1.3.4
pip==21.0.1
pytz==2021.1
setuptools==54.2.0
sqlparse==0.4.1
